Factor out and test game evaluation
